Improved extension features (#2)
* Update: Site logging

Logging all the company names and sites that the extension comes across

* Update: Application form handling sign in

Handling application sign in and create account

* Update: Connecting to nativeHost

establishing connection to app
Adding a test-app (slimmer for testing)

* Chore: Reviewing and commenting the site logger

Ensuring the site logging functionality works well

* Chore: Reviewing and commenting the account autofiller

* Update: Adding HomeLink Icon

* Update: Adding target button to the page

Testing how to add the target button to the page

* Update: Adding Icon and text

Adding the icon and the text and blending the styles in

* Update: Button blend

Ensuring the buttons have the right color and style and all

* Adding myWorkday Home page

Adding the button link to the home page

* Update: Adding Account helper buttons

Adding signIn and Register helper buttons

* Update: Making the button last element in nav

Ensuing my button is always last element in nav

* Update background.js

* Update: Restablishing connection with nativehost

Fixing the connection with nativehost (still a persistent connection issue though).
Tailoring the logic to what's needed.

* Update: testing protocol update

Also trying getData (scraping) and displaying in the home page

* Update: Collecting and sending data with nativehost

Managing received data from nativehost for scraping and sending data to nativehost for addsite